+++ src/netbots_server.py
@@ -1,4 +1,5 @@
 import argparse
+import sys
 import signal
 import time
 import random
@@ -22,7 +23,7 @@ class SrvData():
     conf = {
         # Static vars (some are settable at start up by server command line switches and then do not change after that.)
         'serverName': "NetBot Server",
-        'serverVersion': "1.2.0",
+        'serverVersion': "1.3.0",
 
         # Game and Tournament
         'botsInGame': 4,  # Number of bots required to join before game can start.
@@ -37,6 +38,7 @@ class SrvData():
         # Number of msgs from a bot that server will respond to each step. Others in Q will be dropped.
         'botMsgsPerStep': 4,
         'allowRejoin': True,  # Allows crashed bots to rejoin game in progress.
+        'noViewers': False,  # if True addViewerRequest messages will be rejected. 
 
         # Sizes
         # Area is a square with each side = arenaSize units (0,0 is bottom left,
@@ -264,7 +266,7 @@ def findOverlapingBots(d, bots):
         if 'health' not in boti or boti['health'] is not 0:
             for j in range(i + 1, len(keys)):
                 botj = bots[keys[j]]
-                if 'health' not in boti or botj['health'] is not 0:
+                if 'health' not in botj or botj['health'] is not 0:
                     if nbmath.distance(boti['x'], boti['y'], botj['x'], botj['y']) <= d.conf['botRadius'] * 2:
                         return [keys[i], keys[j]]
 
@@ -770,6 +772,8 @@ def main():
                         default=0, help='How many jam zones does the arena have.')
     parser.add_argument('-startperms', dest='startPermutations', action='store_true',
                         default=False, help='Use all permutations of each set of random start locations.')
+    parser.add_argument('-noviewers', dest='noViewers', action='store_true',
+                        default=False, help='Do not allow viewers.')
     parser.add_argument('-debug', dest='debug', action='store_true',
                         default=False, help='Print DEBUG level log messages.')
     parser.add_argument('-verbose', dest='verbose', action='store_true',
@@ -797,12 +801,13 @@ def main():
     d.conf['obstacles'] = mkObstacles(d, args.obstacles)
     d.conf['jamZones'] = mkJamZones(d, args.jamZones)
     d.conf['startPermutations'] = args.startPermutations
+    d.conf['noViewers'] = args.noViewers
     
     mkStartLocations(d)
 
     log("Server Name: " + d.conf['serverName'])
     log("Server Version: " + d.conf['serverVersion'])
-
+    log("Argument List:" + str(sys.argv))
     log("Server Configuration: " + str(d.conf), "VERBOSE")
 
     try: