+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2019-05-15
+### Added
+- Server option -noviewers which disables viewers.
+- Server prints command line options on startup.
+
+### Changed
+- netbots_math.contains() can now scan a full circle (0 to 0 radians)
+- Robot colors in viewer have been improved.
+- Minor documentation fixes.
+
 ## [1.2.0] - 2019-05-06
 ### Added
 - Ability to use all permutations of each start set of start locations. Enable with -startperms command line option. When enabled, each set of start locations will be used multiple times with each permutation of robots to positions. 
@@ -50,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.9.0] - 2019-04-16
 - Beta release.
 
-[Unreleased]: https://github.com/dbakewel/netbots/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/dbakewel/netbots/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/dbakewel/netbots/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/dbakewel/netbots/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/dbakewel/netbots/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/dbakewel/netbots/compare/1.0.1...1.1.0